requirements.txt
beautifulsoup4==4.10.0
Flask==2.1.1
line-bot-sdk==1.21.0
requests==2.26.0
spacy==3.2.1
